+++ app.js
@@ -29,7 +29,7 @@ async function init() {
     return {
       time: Date.now(),
       bingo:  bingoRouter.graphqlResolver(request, response)
-    }
+    };
   };
   // database setup
   let pgPool = new pg.Pool({
@@ -91,7 +91,7 @@ async function init() {
   });
 
 // error handler
-  app.use(function(err, req, res, next) {
+  app.use(function(err, req, res) {
     // set locals, only providing error in development
     res.locals.message = err.message;
     res.locals.error = req.app.get('env') === 'development' ? err : {};